feat(front): adds Timeline component (#1059)
